RW-2795 move highlight function to utils
