+++ ui/src/app/components/help-sidebar.tsx
@@ -5,7 +5,7 @@ import * as React from 'react';
 import {SidebarContent} from 'app/cohort-review/sidebar-content/sidebar-content.component';
 import {ClrIcon} from 'app/components/icons';
 import colors, {colorWithWhiteness} from 'app/styles/colors';
-import {reactStyles, ReactWrapperBase} from 'app/utils';
+import {highlightSearchTerm, reactStyles, ReactWrapperBase} from 'app/utils';
 
 const sidebarContent = require('assets/json/help-sidebar.json');
 
@@ -177,29 +177,6 @@ export class HelpSidebar extends React.Component<Props, State> {
     this.setState({filteredContent});
   }
 
-  highlightSearchTerm(stringToHighlight: string) {
-    const {searchTerm} = this.state;
-    if (searchTerm === '' || searchTerm.length < 3) {
-      return stringToHighlight;
-    }
-    const words: string[] = [];
-    const matchString = new RegExp(searchTerm.trim(), 'i');
-    const matches = stringToHighlight.match(new RegExp(matchString, 'gi'));
-    const splits = stringToHighlight.split(new RegExp(matchString, 'gi'));
-    if (matches) {
-      for (let i = 0; i < matches.length; i++) {
-        words.push(splits[i], matches[i]);
-      }
-      words.push(splits[splits.length - 1]);
-    } else {
-      words.push(splits[0]);
-    }
-    return words.map((word, w) => <span key={w}
-      style={matchString.test(word.toLowerCase()) ? styles.highlighted : {}}>
-        {word}
-      </span>);
-  }
-
   onIconClick(icon: string) {
     const {activeIcon, sidebarOpen} = this.state;
     const newSidebarOpen = !(icon === activeIcon && sidebarOpen);
@@ -272,14 +249,14 @@ export class HelpSidebar extends React.Component<Props, State> {
             </div>
             {displayContent.length > 0
               ? displayContent.map((section, s) => <div key={s}>
-                <h3 style={styles.sectionTitle}>{this.highlightSearchTerm(section.title)}</h3>
+                <h3 style={styles.sectionTitle}>{highlightSearchTerm(searchTerm, section.title)}</h3>
                 {section.content.map((content, c) => {
                   return typeof content === 'string'
-                    ? <p key={c} style={styles.contentItem}>{this.highlightSearchTerm(content)}</p>
+                    ? <p key={c} style={styles.contentItem}>{highlightSearchTerm(searchTerm, content)}</p>
                     : <div key={c}>
-                      <h4 style={styles.contentTitle}>{this.highlightSearchTerm(content.title)}</h4>
+                      <h4 style={styles.contentTitle}>{highlightSearchTerm(searchTerm, content.title)}</h4>
                       {content.content.map((item, i) =>
-                        <p key={i} style={styles.contentItem}>{this.highlightSearchTerm(item)}</p>)
+                        <p key={i} style={styles.contentItem}>{highlightSearchTerm(searchTerm, item)}</p>)
                       }
                     </div>;
                 })}
